+++ lib/src/event_bus_identifier_base.dart
@@ -1,0 +1,6 @@
+abstract class EventBusIdentifierBase<T> {
+  String get eventName;
+  Type get type;
+  int get key;
+  int get historySize;
+}