+++ lib/src/sub_event_identifier.dart
@@ -1,23 +1,30 @@
 import 'package:event_bus_riverpod/src/event_bus_identifier.dart';
+import 'package:event_bus_riverpod/src/event_bus_identifier_base.dart';
 import 'package:event_bus_riverpod/src/event_bus_provider.dart';
 
-class SubEventIdentifier<T> {
+class SubEventIdentifier<T> extends EventBusIdentifierBase<T> {
   final EventBusIdentifier<T> parentEvent;
-  final String subEventName;
+  @override
+  final String eventName;
   final ListenerWhere<T> where;
+  @override
   final int historySize;
 
   SubEventIdentifier(
-    this.subEventName, {
+    this.eventName, {
     required this.parentEvent,
     required this.where,
     this.historySize = 0,
   }) : assert(historySize >= 0);
 
+  @override
+  Type get type => T;
+
   int? _key;
-  int get key => _key ??= Object.hash(parentEvent.key, subEventName);
+  @override
+  int get key => _key ??= Object.hash(parentEvent.key, eventName);
 
   @override
   String toString() =>
-      'SubEventIdentifier(${parentEvent.eventName}.$subEventName, $T)';
+      'SubEventIdentifier(${parentEvent.eventName}.$eventName, $T)';
 }